Little bit of Value cleanup
